LPS-169362 Move util class to API
