+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/action/executor/AddObjectEntryObjectActionExecutorImpl.java
@@ -17,7 +17,7 @@ package com.liferay.object.internal.action.executor;
 import com.liferay.dynamic.data.mapping.expression.DDMExpressionFactory;
 import com.liferay.object.action.executor.ObjectActionExecutor;
 import com.liferay.object.constants.ObjectActionExecutorConstants;
-import com.liferay.object.internal.action.util.ObjectEntryVariablesUtil;
+import com.liferay.object.entry.util.ObjectEntryVariablesUtil;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectRelationship;
 import com.liferay.object.rest.dto.v1_0.ObjectEntry;
